requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.55.8
fsspec==2024.12.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics @ git+https://github.com/ultralytics/ultralytics.git@9d282ddeee5a5e4435fca7170e91e6ea27bf1911
ultralytics-thop==2.0.14
urllib3==2.3.0
